ai-search-demo/requirements.txt
streamlit
cohere
azure-ai-documentintelligence
pandas
PyMuPDF
markdown2
numpy
Pillow
requests
python-dotenv
azure-ai-inference
azure-search-documents
PyMuPDF
autogen-core
autogen-agentchat
autogen-ext[openai]
autogen-ext[azure]
aiohttp